jpegblur.cc (reduceMaxMCU): fix blurring of images with chroma subsampling.
There was a bug in the logic when reducing the mask to MCU and there
was downsampling of a component in the rows dimension.  It lead to
blurring elsewhere in the image but because it was only introduced in
the downsampled components, i.e., typically in the CbCr components,
and not on the Y component, it was barely perceptible.  It could be
seen by displaying a mask of differences between the original and
blurred image.